pyproject.toml
[project]
name = "lifelogger"
version = "0.1.0"
description = "24小时生活录音 + 声纹识别 + 自动转写 → Vault"
requires-python = ">=3.10"
dependencies = [
    "faster-whisper>=1.0.0",
    "pyannote.audio>=3.1.0",
    "speechbrain>=1.0.0",
    "pydub>=0.25.0",
    "numpy>=1.24.0",
    "scipy>=1.11.0",
    "webrtcvad>=2.0.10",
    "python-dotenv>=1.0.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
    "torchaudio>=2.0.0",
]

[project.scripts]
lifelogger = "main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["recorder", "diarizer", "transcriber", "vault", "pipeline"]

[tool.ruff]
line-length = 100
target-version = "py310"
